+++ src/main/java/janggi/domain/board/Direction.java
@@ -1,4 +1,4 @@
-package janggi.domain;
+package janggi.domain.board;
 
 import java.util.Collections;
 import java.util.List;
