+++ src/main/java/janggi/domain/game/Side.java
@@ -1,4 +1,4 @@
-package janggi.domain;
+package janggi.domain.game;
 
 public enum Side {
     CHO("초"),
