+++ src/main/java/janggi/domain/piece/PieceVO.java
@@ -1,4 +1,6 @@
-package janggi.domain;
+package janggi.domain.piece;
+
+import janggi.domain.game.Side;
 
 public record PieceVO(Side side, PieceType type, String pieceNumber) {
