+++ src/main/java/janggi/dto/PlayerDTO.java
@@ -1,6 +1,6 @@
 package janggi.dto;
 
-import janggi.domain.Side;
+import janggi.domain.game.Side;
 
 public record PlayerDTO(String name, Side side) {
 }
